pyproject.toml
[project]
name = "mamimind"
version = "0.1.0"
description = "Agentic AI Legal Document Intelligence System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",

    # GCP
    "google-cloud-storage>=2.14.0",
    "google-cloud-firestore>=2.14.0",
    "google-cloud-pubsub>=2.19.0",
    "google-cloud-logging>=3.9.0",
    "google-cloud-aiplatform>=1.40.0",

    # AI/ML
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "sentence-transformers>=2.3.0",

    # Vector DB
    "qdrant-client>=1.7.0",

    # OCR
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0",

    # Document Processing
    "pypdf>=4.0.0",
    "pymupdf>=1.23.0",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",

    # Utils
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=html --cov-report=term"
